+++ src/EVault/shared/LiquidityUtils.sol
@@ -11,11 +11,12 @@ abstract contract LiquidityUtils is BorrowUtils, LTVUtils {
     using TypesLib for uint256;
 
     // Calculate the value of liabilities, and the liquidation or borrowing LTV adjusted collateral value.
-    function calculateLiquidity(MarketCache memory marketCache, address account, address[] memory collaterals, LTVType ltvType)
-        internal
-        view
-        returns (uint256 collateralValue, uint256 liabilityValue)
-    {
+    function calculateLiquidity(
+        MarketCache memory marketCache,
+        address account,
+        address[] memory collaterals,
+        LTVType ltvType
+    ) internal view returns (uint256 collateralValue, uint256 liabilityValue) {
         validateOracle(marketCache);
 
         for (uint256 i; i < collaterals.length; ++i) {
@@ -38,7 +39,7 @@ abstract contract LiquidityUtils is BorrowUtils, LTVUtils {
         uint256 liabilityValue = getLiabilityValue(marketCache, account, owed);
         if (liabilityValue == 0) return;
 
-        uint collateralValue;
+        uint256 collateralValue;
         for (uint256 i; i < collaterals.length; ++i) {
             collateralValue += getCollateralValue(marketCache, account, collaterals[i], LTVType.BORROWING);
             if (collateralValue >= liabilityValue) return;
@@ -47,16 +48,11 @@ abstract contract LiquidityUtils is BorrowUtils, LTVUtils {
         revert E_AccountLiquidity();
     }
 
-
     // Check if the account has no collateral balance left, used for debt socialization
     // If LTV is zero, the collateral can still be liquidated.
     // If the price of collateral is zero, liquidations are not executed, so the check won't be performed.
     // If there is no collateral balance at all, then debt socialization can happen.
-    function checkNoCollateral(address account, address[] memory collaterals)
-        internal
-        view
-        returns (bool)
-    {
+    function checkNoCollateral(address account, address[] memory collaterals) internal view returns (bool) {
         for (uint256 i; i < collaterals.length; ++i) {
             address collateral = collaterals[i];
 
@@ -69,9 +65,11 @@ abstract contract LiquidityUtils is BorrowUtils, LTVUtils {
         return true;
     }
 
-
-
-    function getLiabilityValue(MarketCache memory marketCache, address account, Owed owed) internal view returns (uint value) {
+    function getLiabilityValue(MarketCache memory marketCache, address account, Owed owed)
+        internal
+        view
+        returns (uint256 value)
+    {
         // update owed with interest accrued
         uint256 owedAssets = getCurrentOwed(marketCache, account, owed).toAssetsUp().toUint();
 
@@ -85,17 +83,21 @@ abstract contract LiquidityUtils is BorrowUtils, LTVUtils {
         }
     }
 
-    function getCollateralValue(MarketCache memory marketCache, address account, address collateral, LTVType ltvType) internal view returns (uint value) {
-            ConfigAmount ltv = getLTV(collateral, ltvType);
-            if (ltv.isZero()) return 0;
+    function getCollateralValue(MarketCache memory marketCache, address account, address collateral, LTVType ltvType)
+        internal
+        view
+        returns (uint256 value)
+    {
+        ConfigAmount ltv = getLTV(collateral, ltvType);
+        if (ltv.isZero()) return 0;
 
-            uint256 balance = IERC20(collateral).balanceOf(account);
-            if (balance == 0) return 0;
+        uint256 balance = IERC20(collateral).balanceOf(account);
+        if (balance == 0) return 0;
 
-            // bid price for collateral
-            (uint256 currentCollateralValue,) = marketCache.oracle.getQuotes(balance, collateral, marketCache.unitOfAccount);
+        // bid price for collateral
+        (uint256 currentCollateralValue,) = marketCache.oracle.getQuotes(balance, collateral, marketCache.unitOfAccount);
 
-            return ltv.mul(currentCollateralValue);
+        return ltv.mul(currentCollateralValue);
     }
 
     function validateOracle(MarketCache memory marketCache) internal pure {
